requirements.txt
Flask==2.3.3
tensorflow==2.13.0
Pillow==10.0.1
numpy==1.24.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2
gunicorn==21.2.0
python-dotenv==1.0.0
